Reorganize SiPixelRawToDigi as SiPixelRawToDigiHeterogeneous using HeterogeneousEDProducer
