+++ EventFilter/SiPixelRawToDigi/plugins/BuildFile.xml
@@ -2,8 +2,11 @@
 <library file="SiPixelDigiToRaw.cc SiPixelRawToDigi.cc" name="EventFilterSiPixelRawToDigiPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
-<library file="SiPixelRawToDigiGPU.cc SiPixelFedCablingMapGPU.cc SiPixelRawToDigiGPUKernel.cu" name="EventFilterSiPixelRawToDigiGPUPlugins">
+<library file="SiPixelRawToDigiHeterogeneous.cc SiPixelFedCablingMapGPU.cc SiPixelRawToDigiGPUKernel.cu" name="EventFilterSiPixelRawToDigiGPUPlugins">
   <use name="cuda"/>
+  <use name="cuda-api-wrappers"/>
   <use name="CalibTracker/SiPixelESProducers"/>
+  <use name="HeterogeneousCore/Producer"/>
+  <use name="HeterogeneousCore/CUDACore"/>
   <flags EDM_PLUGIN="1"/>
 </library>
