+++ EventFilter/SiPixelRawToDigi/python/SiPixelRawToDigi_cfi.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 import EventFilter.SiPixelRawToDigi.siPixelRawToDigi_cfi
-import EventFilter.SiPixelRawToDigi.siPixelRawToDigiGPU_cfi
+import EventFilter.SiPixelRawToDigi.siPixelRawToDigiHeterogeneous_cfi
 
 siPixelDigis = EventFilter.SiPixelRawToDigi.siPixelRawToDigi_cfi.siPixelRawToDigi.clone()
 siPixelDigis.Timing = cms.untracked.bool(False)
@@ -19,30 +19,24 @@ siPixelDigis.UsePhase1 = cms.bool(False)
 siPixelDigis.Regions = cms.PSet( ) 
 siPixelDigis.CablingMapLabel = cms.string("")
 
-siPixelDigisGPU = EventFilter.SiPixelRawToDigi.siPixelRawToDigiGPU_cfi.siPixelRawToDigiGPU.clone()
-siPixelDigisGPU.Timing = cms.untracked.bool(False)
-siPixelDigisGPU.IncludeErrors = cms.bool(True)
-siPixelDigisGPU.InputLabel = cms.InputTag("rawDataCollector")
-siPixelDigisGPU.UseQualityInfo = cms.bool(False)
+_siPixelDigisHeterogeneous = EventFilter.SiPixelRawToDigi.siPixelRawToDigiHeterogeneous_cfi.siPixelRawToDigiHeterogeneous.clone()
+_siPixelDigisHeterogeneous.IncludeErrors = cms.bool(True)
+_siPixelDigisHeterogeneous.InputLabel = cms.InputTag("rawDataCollector")
+_siPixelDigisHeterogeneous.UseQualityInfo = cms.bool(False)
 ## ErrorList: list of error codes used by tracking to invalidate modules
-siPixelDigisGPU.ErrorList = cms.vint32(29)
+_siPixelDigisHeterogeneous.ErrorList = cms.vint32(29)
 ## UserErrorList: list of error codes used by Pixel experts for investigation
-siPixelDigisGPU.UserErrorList = cms.vint32(40)
+_siPixelDigisHeterogeneous.UserErrorList = cms.vint32(40)
 ##  Use pilot blades
-siPixelDigisGPU.UsePilotBlade = cms.bool(False)
+_siPixelDigisHeterogeneous.UsePilotBlade = cms.bool(False)
 ##  Use phase1
-siPixelDigisGPU.UsePhase1 = cms.bool(False)
+_siPixelDigisHeterogeneous.UsePhase1 = cms.bool(False)
 ## Empty Regions PSet means complete unpacking
-siPixelDigisGPU.Regions = cms.PSet( )
-siPixelDigisGPU.CablingMapLabel = cms.string("")
-siPixelDigisGPU.enableErrorDebug = cms.bool(False)
+_siPixelDigisHeterogeneous.Regions = cms.PSet( )
+_siPixelDigisHeterogeneous.CablingMapLabel = cms.string("")
+
+from Configuration.ProcessModifiers.gpu_cff import gpu
+gpu.toReplaceWith(siPixelDigis, _siPixelDigisHeterogeneous)
 
 from Configuration.Eras.Modifier_phase1Pixel_cff import phase1Pixel
 phase1Pixel.toModify(siPixelDigis, UsePhase1=True)
-phase1Pixel.toModify(siPixelDigisGPU, UsePhase1=True)
-
-# In principle I would like to hide the name 'siPixelDigisGPU', but it
-# is used in test/runRawToDigi_GPU_phase1.py which I also don't want
-# to break
-from Configuration.ProcessModifiers.gpu_cff import gpu
-gpu.toReplaceWith(siPixelDigis, siPixelDigisGPU)
